removed warnings related context in transport/*

+++ xlators/transport/ib-sdp/src/ib-sdp.c
@@ -912,7 +912,7 @@ brick_read (struct xlator *xl,
   struct brick_private *priv = xl->private;
   dict_t request = STATIC_DICT;
   dict_t reply = STATIC_DICT;
-  int fd;
+  long fd;
   if (priv->is_debug) {
     FUNCTION_CALLED;
   }
@@ -922,7 +922,7 @@ brick_read (struct xlator *xl,
   if (tmp == NULL) {
     return -1;
   }
-  fd = (int)tmp->context;
+  fd = (long)tmp->context;
 
   {
     //    data_t *prefilled = bin_to_data (buf, size);
@@ -966,7 +966,7 @@ brick_write (struct xlator *xl,
   struct brick_private *priv = xl->private;
   dict_t request = STATIC_DICT;
   dict_t reply = STATIC_DICT;
-  int fd;
+  long fd;
   if (priv->is_debug) {
     FUNCTION_CALLED;
   }
@@ -976,7 +976,7 @@ brick_write (struct xlator *xl,
   if (tmp == NULL) {
     return -1;
   } 
-  fd = (int)tmp->context;
+  fd = (long)tmp->context;
 
   {
     dict_set (&request, "PATH", str_to_data ((char *)path));
@@ -1092,7 +1092,7 @@ brick_flush (struct xlator *xl,
   struct brick_private *priv = xl->private;
   dict_t request = STATIC_DICT;
   dict_t reply = STATIC_DICT;
-  int fd;
+  long fd;
   if (priv->is_debug) {
     FUNCTION_CALLED;
   }
@@ -1102,7 +1102,7 @@ brick_flush (struct xlator *xl,
   if (tmp == NULL) {
     return -1;
   }
-  fd = (int)tmp->context;
+  fd = (long)tmp->context;
 
   {
     dict_set (&request, "PATH", str_to_data ((char *)path));
@@ -1138,7 +1138,7 @@ brick_release (struct xlator *xl,
   struct brick_private *priv = xl->private;
   dict_t request = STATIC_DICT;
   dict_t reply = STATIC_DICT;
-  int fd;
+  long fd;
   if (priv->is_debug) {
     FUNCTION_CALLED;
   }
@@ -1148,7 +1148,7 @@ brick_release (struct xlator *xl,
   if (tmp == NULL) {
     return -1;
   } 
-  fd = (int)tmp->context;
+  fd = (long)tmp->context;
 
   {
     dict_set (&request, "PATH", str_to_data ((char *)path));
@@ -1197,7 +1197,7 @@ brick_fsync (struct xlator *xl,
   struct brick_private *priv = xl->private;
   dict_t request = STATIC_DICT;
   dict_t reply = STATIC_DICT;
-  int fd;
+  long fd;
   if (priv->is_debug) {
     FUNCTION_CALLED;
   }
@@ -1207,7 +1207,7 @@ brick_fsync (struct xlator *xl,
   if (tmp == NULL) {
     return -1;
   }
-  fd = (int)tmp->context;
+  fd = (long)tmp->context;
 
   {
     dict_set (&request, "PATH", str_to_data ((char *)path));
@@ -1430,7 +1430,7 @@ brick_opendir (struct xlator *xl,
 
   {
     dict_set (&request, "PATH", str_to_data ((char *)path));
-    dict_set (&request, "FD", int_to_data ((int)tmp->context));
+    dict_set (&request, "FD", int_to_data ((long)tmp->context));
   }
 
   ret = fops_xfer (priv, OP_OPENDIR, &request, &reply);
@@ -1630,7 +1630,7 @@ brick_ftruncate (struct xlator *xl,
   struct brick_private *priv = xl->private;
   dict_t request = STATIC_DICT;
   dict_t reply = STATIC_DICT;
-  int fd;
+  long fd;
   if (priv->is_debug) {
     FUNCTION_CALLED;
   }
@@ -1640,7 +1640,7 @@ brick_ftruncate (struct xlator *xl,
   if (tmp == NULL) {
     return -1;
   } 
-  fd = (int)tmp->context;
+  fd = (long)tmp->context;
 
   {
     dict_set (&request, "PATH", str_to_data ((char *)path));
@@ -1691,7 +1691,7 @@ brick_fgetattr (struct xlator *xl,
 
   {
     dict_set (&request, "PATH", str_to_data ((char *)path));
-    dict_set (&request, "FD", int_to_data ((int)tmp->context));
+    dict_set (&request, "FD", int_to_data ((long)tmp->context));
   }
 
   ret = fops_xfer (priv, OP_FGETATTR, &request, &reply);
